Fixed interpolation issue and threading problem that leaves last section of the window unrendered
The interpolation problem turns out to be a mix of a bunch of issues, first of all the off by 1 value for the y, second of all the get_pixel_color function never set the rgb values so we we're reading empty black colors the whole time, along with some other minor changes it works now, but the interpolation has caused some artifacts in the final render, I assume its related to the threshold since the cdiff returns are either 0 or extremely small around 1-3. so it almost never supersamples.